Fix context handling
The original implementation cached a context.Context object, which is not the right approach.
We can get away with it in asertx because the lifetime of the program is limited to a single operation that's performed
entirely under one context. But that doesn't work when implementing servers.

Instead, we now wrap request context in order to include tenant id metadata.